chore(globs): update test js globs and add negates
Update the `source.test` glob collection to be an object, with a `js`
property containing glob patterns for JavaScript specs.

Breaks:
- the `globs.source.test` glob, previously a collection of glob patterns,
will be expected to be an object, with a `js` property containing the
collection of globs for `js` spec files. The relevant update from
spell-webapp is required for this to work correctly.